Changed how objc method are selected in the CLOS interface
Before this patch for each ObjC method a corresponding CLOS method was
generated and eventually cached on a file.

Now it is cached just the generic function (reducing the size of cache
of about 80%), and we specialize on compute-discriminating-function.
The specialized compute-discriminating-function returns a general
function to call ObjC method.

Exactly it returns a function that sets the installed function to a
 function that calls the right ObjC method, avoiding to compile n
 thousands function each time a framework is loaded.

darcs-hash:d5bd6d0c6ce352d3b88f831014c46e9546b0692b