New framework management
Now framework definitions are not more generated on the fly and
eventually cached.

They are generated the first time for the standard framework and
compiled for faster loading with import-framework. A compile-framework
is provided to make the user able to build the framework definition
for its custom frameworks.

The tests involving the foundation framework have been updated too.

darcs-hash:95f1c3af1c3ff2538d8db6053e8d07983f02f230